Licenta/requirements.txt
pygame~=2.5.2

torch~=2.3.0
numpy~=1.26.4
torchvision~=0.18.0
pillow~=10.3.0
matplotlib~=3.10.0